Added logic for listing cars method.

+++ src/main/java/org/example/biluthyrningssystem/services/CarService.java
@@ -26,32 +26,36 @@ public class CarService implements CarServiceInterface {
     public CarService(CarRepository carRepository) {
         this.carRepository = carRepository;
     }
-
+//LISTAR INTE BILAR SOM ÄR NYLIGEN TILLAGDA OCH HAR BOOKED DATES = NULL
     @Override
     public List<CarDTO> getAvailableCars() {
-        List<CarDTO> carDTO = new ArrayList<>();
+        List<CarDTO> carDTOList = new ArrayList<>();
         LocalDate startDate;
         LocalDate endDate;
         for (Car car : carRepository.findAll()) {
             if (!car.isInService()) {
                 Map<LocalDate, LocalDate> dates = new HashMap<>();
+                if(car.getOrders() != null) {
+                    for (Order order : car.getOrders()) {
+                        if (order.isActive()) {
+                            startDate = order.getStartDate();
+                            endDate = order.getEndDate();
 
-                for (Order order : car.getOrders()) {
-                    if (order.isActive()) {
-                        startDate = order.getStartDate();
-                        endDate = order.getEndDate();
-
-                        dates.put(startDate, endDate);
+                            dates.put(startDate, endDate);
+                        }
+                        carDTOList.add(new CarDTO(car, dates));
                     }
-                    carDTO.add(new CarDTO(car, dates));
+                }
+                else {
+                    carDTOList.add(new CarDTO(car));
                 }
             }
         }
         Comparator<CarDTO> priceComparator = (car1, car2) -> (int) (car1.getPricePerDay() - car2.getPricePerDay());
-        carDTO.sort(priceComparator);
-        return carDTO;
+        carDTOList.sort(priceComparator);
+        return carDTOList;
     }
-
+//LISTAR JUST NU BARA BILAR SOM INTE ÄR PÅ SERVICE!
     @Override
     public List<Car> adminGetAvailableCars() {
         List<Car> cars = new ArrayList<>();
@@ -80,8 +84,10 @@ public class CarService implements CarServiceInterface {
         if(isValidCar(car)) {
             carRepository.save(car);
             USER_LOGGER.info("Admin added a car with plate number {}.", car.getPlateNumber());
+            return "Car added";
         }
-        return "Car added";
+        USER_LOGGER.warn("Admin tried to add a car with invalid data.");
+        throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Missing some data.");
     }
 
     @Override
@@ -92,6 +98,8 @@ public class CarService implements CarServiceInterface {
                 USER_LOGGER.info("Admin updated car with plate number {}.", car.getPlateNumber());
                 return "Car updated";
             }
+            USER_LOGGER.warn("Admin tried to update a car with invalid data.");
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Missing some data.");
         }
         USER_LOGGER.warn("Admin tried to update a car that does not exist.");
         throw new ResourceNotFoundException("Car", "id", car.getId());
@@ -108,6 +116,11 @@ public class CarService implements CarServiceInterface {
                     throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Car is rented by a customer and can not be removed");
                 }
             }
+            if(carToRemove.isInService()) {
+                USER_LOGGER.warn("Admin tried to remove a car that is in service");
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Car is in service and can not be removed");
+            }
+
             carRepository.deleteById(id);
             USER_LOGGER.info("Admin removed car with plate number {}.", carToRemove.getPlateNumber());
             return "Car removed";
@@ -117,18 +130,22 @@ public class CarService implements CarServiceInterface {
     }
 
     private boolean isValidCar(Car car) {
-        if(car.getPricePerDay() <= 0) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid pricePerDay");
-        }
-        else if(car.getBrand() == null || car.getBrand().isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "brand is missing");
-        }
-        else if(car.getModel() == null || car.getModel().isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "model is missing");
-        }
-        else if(car.getPlateNumber() == null || car.getPlateNumber().isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "plateNumber is missing");
-        }
-        return true;
+        return !(car.getPricePerDay() <= 0) &&
+                car.getBrand() != null && !car.getBrand().isEmpty() &&
+                car.getModel() != null && !car.getModel().isEmpty() &&
+                car.getPlateNumber() != null && !car.getPlateNumber().isEmpty();
     }
+//        if(car.getPricePerDay() <= 0) { //ÄNDRA DENNA METOD
+//            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Invalid pricePerDay");
+//        }
+//        else if(car.getBrand() == null || car.getBrand().isEmpty()) {
+//            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "brand is missing");
+//        }
+//        else if(car.getModel() == null || car.getModel().isEmpty()) {
+//            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "model is missing");
+//        }
+//        else if(car.getPlateNumber() == null || car.getPlateNumber().isEmpty()) {
+//            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "plateNumber is missing");
+//        }
+//        return true;
 }
